requirements.txt
algoliasearch==1.12.0
algoliasearch-django==1.4.1
asn1crypto==0.22.0
beautifulsoup4==4.6.0
boto3==1.4.4
botocore==1.5.83
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
cryptography==1.9
dj-database-url==0.4.2
Django==1.11.4
django-modelcluster==3.1
django-storages==1.6.3
django-taggit==0.22.1
django-treebeard==4.1.2
djangorestframework==3.6.3
docutils==0.13.1
gunicorn==19.7.1
html5lib==0.999999999
idna==2.6
jmespath==0.9.3
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.1
pycparser==2.18
pyOpenSSL==17.1.0
python-dateutil==2.6.1
python-dotenv==0.7.1
pytz==2017.2
requests==2.18.4
s3transfer==0.1.10
six==1.10.0
Unidecode==0.4.21
urllib3==1.22
wagtail==1.12.2
wagtailgeowidget==3.1.3
webencodings==0.5.1
whitenoise==3.3.0
Willow==0.4
